--- SixthSense/UICore/Sources/AppAlertViewController.swift
@@ -1,234 +1,0 @@
-//
-//  AppAlertViewController.swift
-//  UICore
-//
-//  Created by Allie Kim on 2022/08/23.
-//
-
-import Then
-import SnapKit
-import UIKit
-import RxSwift
-import RxCocoa
-import DesignSystem
-
-public typealias AlertActionCompletion = ((AlertAction) -> Void)
-
-public struct AlertAction {
-
-    public enum Style {
-        case positive // 오른쪽
-        case negative // 왼쪽
-
-        var attributes: [NSAttributedString.Key: Any] {
-            switch self {
-            case .positive:
-                return [.font: AppFont.body2Bold, .foregroundColor: UIColor.white]
-            case .negative:
-                return [.font: AppFont.body2Bold, .foregroundColor: AppColor.systemGray500]
-            }
-        }
-
-        var background: UIColor {
-            switch self {
-            case .positive:
-                return AppColor.main
-            case .negative:
-                return .white
-            }
-        }
-
-        var borderColor: CGColor {
-            switch self {
-            case .positive:
-                return AppColor.main.cgColor
-            case .negative:
-                return AppColor.systemGray300.cgColor
-            }
-        }
-    }
-
-    var title: String
-    var style: AlertAction.Style
-
-    public static func action(title: String, style: AlertAction.Style) -> AlertAction {
-        return AlertAction(title: title, style: style)
-    }
-
-    private init(title: String, style: AlertAction.Style) {
-        self.title = title
-        self.style = style
-    }
-}
-
-public final class AppAlertViewController: UIViewController {
-
-    private enum Constants {
-
-        enum Inset {
-            static let `default` = 20.0
-            static let labelStackView = 40.0
-        }
-
-        enum Height {
-            static let button = 50.0
-        }
-
-        enum Width {
-            static let button = 142.0
-        }
-        
-        enum Radius {
-            static let alertView = 15.0
-            static let button = 10.0
-        }
-    }
-
-    private let alertTitle: String
-    private let message: String
-
-    private let alertView = UIView().then {
-        $0.backgroundColor = .white
-        $0.layer.cornerRadius = Constants.Radius.alertView
-    }
-
-    private let labelStackView = UIStackView().then {
-        $0.axis = .vertical
-        $0.spacing = 10
-        $0.alignment = .center
-        $0.distribution = .fillEqually
-    }
-
-    private let titleLabel = UILabel().then {
-        $0.textAlignment = .center
-        $0.numberOfLines = 1
-        $0.font = AppFont.body2Bold
-    }
-
-    private let messageLabel = UILabel().then {
-        $0.textAlignment = .center
-        $0.numberOfLines = 2
-        $0.font = AppFont.body2
-    }
-
-    private let buttonStackView = UIStackView().then {
-        $0.axis = .horizontal
-        $0.spacing = 11
-        $0.distribution = .fillEqually
-    }
-
-    private let disposeBag = DisposeBag()
-
-    init(alertTitle: String, message: String) {
-        self.alertTitle = alertTitle
-        self.message = message
-        super.init(nibName: nil, bundle: nil)
-        modalPresentationStyle = .overFullScreen
-        modalTransitionStyle = .crossDissolve
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    public override func viewDidLoad() {
-        super.viewDidLoad()
-        configureUI()
-        configureLayout()
-    }
-
-    private func configureUI() {
-        view.backgroundColor = .black.withAlphaComponent(0.5)
-        view.addSubview(alertView)
-        alertView.addSubviews(labelStackView, buttonStackView)
-        labelStackView.addArrangedSubviews(titleLabel, messageLabel)
-
-        titleLabel.text = alertTitle
-        messageLabel.text = message
-    }
-
-    private func configureLayout() {
-        alertView.snp.makeConstraints {
-            $0.left.right.equalToSuperview().inset(Constants.Inset.default)
-            $0.center.equalToSuperview()
-        }
-
-        labelStackView.snp.makeConstraints {
-            $0.left.right.equalToSuperview().inset(Constants.Inset.default)
-            $0.top.equalToSuperview().offset(Constants.Inset.labelStackView)
-            $0.bottom.equalTo(buttonStackView.snp.top).offset(-Constants.Inset.default)
-        }
-
-        buttonStackView.snp.makeConstraints {
-            $0.left.right.equalToSuperview().inset(Constants.Inset.default)
-            $0.bottom.equalToSuperview().offset(-Constants.Inset.default)
-            $0.height.equalTo(Constants.Height.button)
-        }
-    }
-
-    private func makeButton(action: AlertAction) -> UIButton {
-        let button = UIButton()
-        button.frame = CGRect(x: 0,
-                              y: 0,
-                              width: Constants.Width.button,
-                              height: Constants.Height.button)
-
-        button.setAttributedTitle(NSAttributedString(string: action.title, attributes: action.style.attributes), for: .normal)
-        button.backgroundColor = action.style.background
-        button.layer.borderColor = action.style.borderColor
-        button.layer.borderWidth = 1
-        button.layer.cornerRadius = Constants.Radius.button
-        button.layer.masksToBounds = true
-        return button
-    }
-
-    private func dismiss() {
-        let transition = CATransition()
-        transition.duration = 0.15
-        transition.timingFunction = CAMediaTimingFunction(name: CAMediaTimingFunctionName.easeInEaseOut)
-        transition.type = CATransitionType.fade
-        self.view.window!.layer.add(transition, forKey: nil)
-        self.dismiss(animated: false)
-    }
-
-    func addButtons(action: AlertAction,
-                    completion: AlertActionCompletion? = nil) {
-        let button = makeButton(action: action)
-        buttonStackView.addArrangedSubview(button)
-        if let completion = completion {
-            button.rx.tap
-                .throttle(.seconds(2),
-                          latest: false,
-                          scheduler: MainScheduler.instance)
-                .withUnretained(self)
-                .bind(onNext: { owner, _ in
-                owner.dismiss()
-                completion(action)
-            })
-                .disposed(by: disposeBag)
-        }
-    }
-}
-
-public extension UIViewController {
-
-    func showAlert(
-        title: String,
-        message: String,
-        actions: [AlertAction]
-    ) -> Observable<AlertAction.Style> {
-        return Observable.create { observer in
-            let alertController = AppAlertViewController(alertTitle: title,
-                                                         message: message)
-
-            actions.forEach { action in
-                alertController.addButtons(action: action) { action in
-                    observer.onNext(action.style)
-                    observer.onCompleted()
-                }
-            }
-            self.present(alertController, animated: true, completion: nil)
-            return Disposables.create { alertController.dismiss(animated: false) }
-        }
-    }
-}
